requirements.txt
selenium==4.26.1 
python-dotenv==1.0.1 
pandas==2.2.3

